pyproject.toml
[project]
name = "stt-typer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyautogui>=0.9.54",
    "pyaudio>=0.2.11",
    "google-genai>=1.19.0",
]
